backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
mediapipe==0.10.18
celery[redis]==5.4.0
python-dotenv==1.0.1
redis==5.0.4
pydantic[email]==2.7.1
requests==2.31.0
pytest==8.2.0
httpx==0.27.0